pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Jefferson Fragoso de Souza", email = "jefferson.fragoso@pgmais.com.br"},
]
dependencies = [
    "django>=4.0.4",
    "djangorestframework>=3.14.0",
    "install>=1.3.5",
    "django-extensions>=3.2.3",
    "dependency-injector>=4.41.0",
]
requires-python = ">=3.10.2"
license = {text = "MIT"}
[project.optional-dependencies]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "autopep8>=1.7.0",
    "pylint>=2.15.3",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-django>=4.5.2",
]

[tool.pdm.scripts]
manage="python ./src/django_app/manage.py"
pep8 = "autopep8 --in-place --recursive ./src"
pylint = "pylint ./src"
runserver = "pdm run manage runserver 0.0.0.0:8000"
pytest = "pytest --ignore __pypackages__"
test_cov = "pdm run test --cov ./src --cov-fail-under 80"
test_cov_html = "pdm run test_cov --cov-report html:./__coverage"
